feat: composer configuration & remove /vendor
add workflow for composer validation
add composer.lock for php7 and php8
install.sh and restore.php will install composer dependancies

+++ install/install.sh
@@ -324,13 +324,20 @@ step_9_jeedom_configuration() {
 step_10_jeedom_installation() {
   echo "---------------------------------------------------------------------"
   echo "${YELLOW}Starting step 10 - Jeedom install${NORMAL}"
+  # install composer
   chmod +x ${WEBSERVER_HOME}/resources/install_composer.sh
   ${WEBSERVER_HOME}/resources/install_composer.sh
-  rm -rf ${WEBSERVER_HOME}/vendor
-  rm -rf ${WEBSERVER_HOME}/composer.lock
   export COMPOSER_ALLOW_SUPERUSER=1
   cd ${WEBSERVER_HOME}
-  composer install --no-ansi --no-dev --no-interaction --no-plugins --no-progress --no-scripts --optimize-autoloader
+
+  PHP_VERSION=$(php -r "echo PHP_VERSION;")
+  if [ $(version $PHP_VERSION) -ge $(version "8.0.0") ]; then
+    echo "PHP version highter than 8.0.0, get composer.php8.lock"
+    rm -rf ${WEBSERVER_HOME}/composer.lock
+    cp ${WEBSERVER_HOME}/composer.php8.lock ${WEBSERVER_HOME}/composer.lock
+  fi
+  composer install --no-progress --no-interaction --no-dev 
+
   mkdir -p /tmp/jeedom
   chmod 777 -R /tmp/jeedom
   chown www-data:www-data -R /tmp/jeedom
